requirements.txt
Django==4.2.16
python-decouple==3.8
psycopg[binary]==3.1.17
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-soft-delete==1.0.15
ipython==8.28.0
